requirements.txt
python-telegram-bot==9.0.0
gunicorn
flask_restful==0.3.6
Flask==0.12.2
flask-jsonpify==1.5.0
requests==2.18.4
emoji==0.5.0

